Develop

+++ routes/web.php
@@ -19,34 +19,34 @@ use App\Http\Controllers\UserController;
 Route::get('/', [BlogController::class, 'index']);
 
 // show create blog form
-Route::get('/blogs/create', [BlogController::class, 'create']);
+Route::get('/blogs/create', [BlogController::class, 'create'])->middleware('auth');
 
 // store blog data
-Route::post('/blogs', [BlogController::class, 'store']);
+Route::post('/blogs', [BlogController::class, 'store'])->middleware('auth')->middleware('auth');
 
 // show edit blog form
-Route::get('/blogs/{blog}/edit', [BlogController::class, 'edit']);
+Route::get('/blogs/{blog}/edit', [BlogController::class, 'edit'])->middleware('auth');
 
 // update blog
-Route::put('/blogs/{blog}', [BlogController::class, 'update']);
+Route::put('/blogs/{blog}', [BlogController::class, 'update'])->middleware('auth');
 
 // delete blog
-Route::delete('/blogs/{blog}', [BlogController::class, 'destroy']);
+Route::delete('/blogs/{blog}', [BlogController::class, 'destroy'])->middleware('auth');
 
 // show single blog
 Route::get('/blogs/{blog}', [BlogController::class, 'show']);
 
 // show registration form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // create new user
 Route::post('/users', [UserController::class, 'store']);
 
 // logout user
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 // show login form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // login user
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
